remarks marika 1
