+++ scripts/replace-common-readme.ts
@@ -1,0 +1,48 @@
+import { resolve  } from 'path';
+import { promises, readFileSync } from 'fs';
+import { createLogger } from '@sap-cloud-sdk/util';
+import { getNonTestLernaModules } from './util';
+import {EOL} from 'os'
+
+const startTagCommonReadme = '<!--sapCloudSdkCommonReadme-->';
+const endTagCommonReadme = '<!--sapCloudSdkCommonReadmeStop-->';
+const startTagLogo = '<!--sapCloudSdkLogo-->';
+const endTagLogo = '<!--sapCloudSdkLogoStop-->'
+
+const logoContent = `<a href="https://sap.com/s4sdk"><img src="https://help.sap.com/doc/2324e9c3b28748a4ae2ad08166d77675/1.0/en-US/logo-with-js.svg" alt="SAP Cloud SDK for JavaScript Logo" height="122.92" width="226.773"/></a>${EOL}`
+
+const logger = createLogger('check-licenses');
+
+function insertCommonContent(oldFileContent: string): string {
+  const genericContent = readFileSync(
+    resolve(__dirname, 'COMMON-README-PART.md'),
+    { encoding: 'utf8' }
+  );
+
+  let newFileContent = replaceContentUsingTags(startTagCommonReadme,endTagCommonReadme,genericContent,oldFileContent)
+  return replaceContentUsingTags(startTagLogo,endTagLogo,logoContent,newFileContent)
+
+}
+
+function replaceContentUsingTags(startTag:string,endTag:string,replacement:string,fileContent:string){
+  return fileContent.replace(new RegExp(`${startTag}(.|\n)*${endTag}`),
+    `${startTag}\n${replacement}${endTag}`)
+}
+
+async function updateReadmeFile(pathModule: string) {
+  const pathReadme = resolve(pathModule, 'README.md');
+  const oldFileContent = await promises.readFile(pathReadme, {
+    encoding: 'utf8'
+  });
+  const newFileContent = insertCommonContent(oldFileContent);
+  await promises.writeFile(pathReadme, newFileContent, { encoding: 'utf8' });
+  logger.info(`File ${pathReadme} finished.`);
+}
+
+async function updateReadmeFiles() {
+  logger.info('Generic content is added to README.md files in packages.');
+  const packages = await getNonTestLernaModules();
+  await Promise.all(packages.map(module => updateReadmeFile(module.location)));
+}
+
+updateReadmeFiles();
